For review

+++ frontend/js/app.js
@@ -76,19 +76,21 @@ app.controller("ContactCtrl", ["$scope", "$http", "$q", "$timeout",
 
             var url = "http://127.0.0.1:8000/api/contact/" + id + "/"; 
 
-            $http.delete(url).then(()=>{
-                $scope.delete = true;
-                $timeout(function(){
-                    $scope.delete = false; 
-                }, 2000);
-                get();
-            }).catch((error)=>{
-                console.log(error);
-                $scope.failed = true;
-                $timeout(function(){
-                    $scope.failed = false; 
-                }, 2000);
-            });
+            if (confirm("Are you sure?")){
+                $http.delete(url).then(()=>{
+                    $scope.delete = true;
+                    $timeout(function(){
+                        $scope.delete = false; 
+                    }, 2000);
+                    get();
+                }).catch((error)=>{
+                    console.log(error);
+                    $scope.failed = true;
+                    $timeout(function(){
+                        $scope.failed = false; 
+                    }, 2000);
+                });
+            }
         };
 
         // loading the contact list
